Update magic.db from Spelltable
NOTE: It seems like there's quite a difference (~5%) between the hashes in the included db vs the new one. I'm guessing it has to do with the fact that the new database generation stuff is done differently. It still seems to work ok though.

This also adds a webserver which can be used to do the comparisons. My project calls this to get the hash results and bundles the results together.